server/requirements.txt
Flask>=2.3.0
firebase-admin>=7.0.0
google-cloud-firestore>=2.10.0
face-recognition>=1.3.0
numpy>=1.24.0
opencv-python>=4.8.0
cloudinary>=1.33.0
Flask-Cors>=3.1.0
python-dotenv>=1.0.0
